+++ src/AppleTv.Movie.Price.Tracker.App/Controllers/MoviesController.cs
@@ -1,19 +1,16 @@
-﻿using System.ComponentModel.DataAnnotations;
-using AppleTv.Movie.Price.Tracker.Data;
+﻿using AppleTv.Movie.Price.Tracker.Domains.Models;
+using AppleTv.Movie.Price.Tracker.Domains.Movies.Commands.TrackMovie;
+using AppleTv.Movie.Price.Tracker.Domains.Movies.Commands.UntrackMovie;
 using AppleTv.Movie.Price.Tracker.Domains.Movies.Models;
 using AppleTv.Movie.Price.Tracker.Domains.Movies.Queries.GetMovies;
 using AppleTv.Movie.Price.Tracker.Domains.Movies.Queries.LookupMovie;
 using AppleTv.Movie.Price.Tracker.Domains.Movies.Queries.SearchMovies;
-using AppleTv.Movie.Price.Tracker.Services;
 using AppleTv.Movie.Price.Tracker.Services.Models;
-using AutoMapper;
 using kr.bbon.AspNetCore;
 using kr.bbon.AspNetCore.Mvc;
-using kr.bbon.Core.Exceptions;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 namespace AppleTv.Movie.Price.Tracker.App.Controllers;
 
@@ -25,12 +22,9 @@ namespace AppleTv.Movie.Price.Tracker.App.Controllers;
 [Produces(Constants.RESPONSE_MEDIA_TYPE)]
 public class MoviesController : ApiControllerBase
 {
-    public MoviesController(IMediator mediator, AppDbContext appDbContext, ITunesSearchService iTunesSearchService, IMapper mapper, ILogger<MoviesController> logger)
+    public MoviesController(IMediator mediator, ILogger<MoviesController> logger)
     {
         this.mediator = mediator;
-        this.appDbContext = appDbContext;
-        this.iTunesSearchService = iTunesSearchService;
-        this.mapper = mapper;
         this.logger = logger;
     }
 
@@ -71,160 +65,24 @@ public class MoviesController : ApiControllerBase
 
     [HttpPost]
     [Route("track")]
-    public async Task<IActionResult> Track([FromBody] TrackModel model)
+    public async Task<ActionResult<MovieModel>> Track([FromBody] TrackMovieCommand command)
     {
-        if (!ModelState.IsValid)
-        {
-            return BadRequest();
-        }
+        var result = await mediator.Send(command);
 
-        var result = await iTunesSearchService.LookupMovieAsync(model.Id, model.Country, model.Language);
-
-        if (result == null)
-        {
-            throw new ApiException(System.Net.HttpStatusCode.NotFound);
-        }
-
-        var movie = await appDbContext.Movies
-            .Where(x => x.TrackId == model.Id && x.CountryCode == model.Country && x.LanguageCode == model.Language)
-            .FirstOrDefaultAsync();
-
-        Guid movieId;
-
-        if (movie != null)
-        {
-            movieId = movie.Id;
-
-            movie.ArtistName = result.ArtistName;
-            movie.ArtworkUrl100 = result.ArtworkUrl100;
-            movie.ArtworkUrl30 = result.ArtworkUrl30;
-            movie.ArtworkUrl60 = result.ArtworkUrl60;
-            movie.ArtworkUrlBase = result.ArtworkUrlBase;
-            movie.CollectionPrice = result.CollectionPrice;
-            movie.CollectionHdPrice = result.CollectionHdPrice;
-            movie.TrackPrice = result.TrackPrice;
-            movie.TrackHdPrice = result.TrackHdPrice;
-            movie.TrackRentalPrice = result.TrackRentalPrice;
-            movie.TrackHdRentalPrice = result.TrackHdRentalPrice;
-
-            await appDbContext.SaveChangesAsync();
-
-            logger.LogInformation("Movie updated: {title} ({trackId};{country})", movie.TrackName, movie.TrackId, movie.Country);
-        }
-        else
-        {
-
-            movie = mapper.Map<Entities.Movie>(result);
-
-            var added = appDbContext.Movies.Add(movie);
-
-            movieId = added.Entity.Id;
-
-            await appDbContext.SaveChangesAsync();
-
-            logger.LogInformation("Movie added: {title} ({trackId};{country})", movie.TrackName, movie.TrackId, movie.Country);
-        }
-
-        if (result.CollectionId > 0)
-        {
-            Guid collectionId;
-
-            var collection = await appDbContext.Collections
-                .Where(x => x.CollectionId == result.CollectionId)
-                .FirstOrDefaultAsync();
-
-            if (collection == null)
-            {
-                collection = new()
-                {
-                    CollectionArtistId = result.CollectionArtistId,
-                    CollectionArtistViewUrl = result.CollectionArtistViewUrl,
-                    CollectionCensoredName = result.CollectionCensoredName,
-                    CollectionHdPrice = result.CollectionHdPrice,
-                    CollectionId = result.CollectionId,
-                    CollectionName = result.CollectionName,
-                    CollectionPrice = result.CollectionPrice,
-                    CollectionViewUrl = result.CollectionViewUrl,
-                };
-
-                var addedCollection = appDbContext.Collections.Add(collection);
-                collectionId = addedCollection.Entity.Id;
-            }
-            else
-            {
-                collection.CollectionArtistId = result.CollectionArtistId;
-                collection.CollectionArtistViewUrl = result.CollectionArtistViewUrl;
-                collection.CollectionCensoredName = result.CollectionCensoredName;
-                collection.CollectionHdPrice = result.CollectionHdPrice;
-                collection.CollectionId = result.CollectionId;
-                collection.CollectionName = result.CollectionName;
-                collection.CollectionPrice = result.CollectionPrice;
-                collection.CollectionViewUrl = result.CollectionViewUrl;
-
-                collectionId = collection.Id;
-            }
-
-            await appDbContext.SaveChangesAsync();
-
-            var collectionMovie = await appDbContext.CollectionMovies
-                .Where(x => x.MovieId == movieId && x.CollectionId == collectionId)
-                .FirstOrDefaultAsync();
-
-            if (collectionMovie == null)
-            {
-                appDbContext.CollectionMovies.Add(new()
-                {
-                    CollectionId = collectionId,
-                    MovieId = movieId,
-                });
-
-                await appDbContext.SaveChangesAsync();
-            }
-        }
-
-        return Accepted();
+        return Accepted(result);
     }
 
     [HttpDelete()]
     [Route("{movieId:guid}/untrack")]
-    public async Task<IActionResult> Untrack([FromRoute] Guid movieId)
+    public async Task<ActionResult<bool>> Untrack([FromRoute] Guid movieId)
     {
-        var movie = await appDbContext.Movies.Where(x => x.Id == movieId).FirstOrDefaultAsync();
+        UntrackMovieCommand command = new() { Id = movieId };
 
-        if (movie == null)
-        {
-            return NotFound();
-        }
+        var result = mediator.Send(command);
 
-        var collectionMovies = appDbContext.CollectionMovies
-            .Where(x => x.MovieId == movie.Id);
-
-        foreach (var item in collectionMovies)
-        {
-            appDbContext.CollectionMovies.Remove(item);
-        }
-
-        appDbContext.Movies.Remove(movie);
-
-        await appDbContext.SaveChangesAsync();
-
-        return Accepted();
+        return Accepted(result);
     }
 
     private readonly IMediator mediator;
-    private readonly AppDbContext appDbContext;
-    private readonly ITunesSearchService iTunesSearchService;
-    private readonly IMapper mapper;
     private readonly ILogger logger;
-}
-
-
-public class TrackModel
-{
-    [Required]
-    public long Id { get; set; }
-    [Required]
-    public string Country { get; set; } = "kr";
-    [Required]
-    public string Language { get; set; } = "ko_kr";
 }
