+++ src/AppleTv.Movie.Price.Tracker.App/Extensions/DependencyInjection/MvcBuilderExtensions.cs
@@ -35,9 +35,8 @@ public static class MvcBuilderExtensions
                 {
                     ContentTypes =
                     {
-                    // using static System.Net.Mime.MediaTypeNames;
                     MediaTypeNames.Application.Json,
-                    MediaTypeNames.Application.Xml
+                    // MediaTypeNames.Application.Xml
                     }
                 };
             };
