+++ src/AppleTv.Movie.Price.Tracker.App/appsettings.json
@@ -24,11 +24,13 @@
       "Name": ""
     }
   },
-  "MoviePriceCollectJob": {
-    "CronSchedule": "0 0 * * * *",
-    "CronTimeZone": null,
-    "RunImmediately": false,
-    "JobName": "MoviePriceCollectJob"
+  "SchedulerJobs": {
+    "MoviePriceCollectJob": {
+      "CronSchedule": "0 0 * * * *",
+      "CronTimeZone": null,
+      "RunImmediately": false,
+      "JobName": "MoviePriceCollectJob"
+    }
   },
   "Logging": {
     "LogLevel": {
