+++ src/AppleTv.Movie.Price.Tracker.Domains/Models/IPrice.cs
@@ -1,0 +1,13 @@
+namespace AppleTv.Movie.Price.Tracker.Domains.Models;
+
+public interface IPrice
+{
+    decimal CollectionPrice { get; }
+    decimal CollectionHdPrice { get; }
+
+    decimal TrackPrice { get; }
+    decimal TrackRentalPrice { get; }
+
+    decimal TrackHdPrice { get; }
+    decimal TrackHdRentalPrice { get; }
+}
